Upload Combat Demo v1.11
Not sure on this naming scheme. It's a bit of an awkward situation,
since all prior builds are only available on a single platform, but I
don't think this is the ideal solution. I might go through and rename
all the builds to something better in the future (would have to update
all links to them, though...).